requirements.txt
Flask>=2.3.2
gpiozero>=2.0.1
Pillow>=10.3.0
